requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
diskcache==5.6.3
fastapi==0.115.12
filelock==3.18.0
flask==3.1.1
flask-socketio==5.5.1
fonttools==4.58.1
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.156
urllib3==2.4.0
uvicorn==0.34.3
werkzeug==3.1.3
wsproto==1.2.0
